fix: Throw InvalidEventException when API Event Refs HttpAPI (#2159)
If a customer miss configures their template by defining an Event of Type
API but Ref an AWS::Serverless::HttpApi, SAM just fails. This results in
a "service error" to customers but SAM should be throwing something helpful
back to customers to help them through the error.

In addition, I updated the test_translator.py tests for error cases to read
all the "error_" templates to test. This allows us not to manually add files
everytime we add a new error case. This caught 3 files that were not being run
and had some misconfigured output. The commit addresses those as well.

Co-authored-by: Jacob Fuss <jfuss@users.noreply.github.com>